chore: optimize headmanage
